refactor: move binfmt_misc functions in main to binfmt.cpp/binfmt.hpp
