+++ CMakeLists.txt
@@ -94,6 +94,7 @@ set(FELIX86_COMMON_SOURCES
     src/felix86/common/strace.cpp
     src/felix86/common/sudo.cpp
     src/felix86/common/utility.cpp
+    src/felix86/common/binfmt.cpp
     src/felix86/common/xsave.cpp
 )
