+++ src/felix86/common/binfmt.hpp
@@ -1,4 +1,6 @@
 #pragma once
 #include <string>
 
+bool detect_binfmt_misc();
+void binfmt_misc(bool is_register);
 bool unregister_binfmt_misc(const std::string& path);
